remove module dependencies
I wasn't sure whether to allow any leveldb passed to the require. What if that handled sublevel modular installment by returning two level database objects... like pass a db to get it back and its sub 

Maybe return an array of the two databses, a sub and its above-level original... 

but I guess it's better to show sublevel use in the example as it was intended, so that its not a dep. 

would it be a good dep? Lack of standards for importing modular levelDBs seems good for experimenting and flexibility but maybe less easy for people IMO